Add servo angles, fix bugs.

+++ depo/hall_probe.cpp
@@ -30,7 +30,7 @@ void HallProbe::update()
 
     if (_changeTime > 0 && millis() >= _changeTime) {
         _changeTime = 0;
-        int period = millis() - _lastEventMs;
+        unsigned long period = millis() - _lastEventMs;
         if ((this->onOccupied != nullptr) && (period >= MIN_EVENT_DELAY)) {
             _lastEventMs = millis();
             onOccupied(this);
